requirements.txt
pandas
numpy
seaborn
matplotlib
scikit-learn
catboost
xgboost
Flask
dill
PyYAML


#-e .